backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.1.0
openai==1.3.0
googlemaps==4.10.0
pytest==7.4.3
httpx==0.25.1
langchain==0.0.350
langgraph==0.0.15
redis==5.0.1
rq==1.15.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
elasticsearch==8.12.1
loguru==0.7.2
jinja2==3.1.3
weasyprint==60.2
aiohttp==3.9.3
beautifulsoup4==4.12.3 